Update ssr.md
Constructor: Although it runs when the component is initialized, the DOM is not yet ready, making it unsuitable for DOM manipulations.

ngAfterViewInit: It runs after the DOM is fully rendered, making it the ideal place to access and manipulate DOM elements. That's why we use this method instead of the constructor.